Fix BooleanConversionFilter bug when nil value at top level
This fixes an issue where if the top-level config hash contains a nil,
we would break out of each_pair, causing the entire execute method to
return nil (and therefore making Chamber return nil for
`Chamber.env.data`).

This might not be the best fix -- I'm not sure what the intention of
breaking if nil was. Depending on the answer to that question, `next if
value.nil?` might be more appropriate.