Bugfix: When running 'chamber secure' only display insecure settings
Previously when you would run `chamber secure`, it would display all settings
which were marked as needing to be secure, whether they were already encrypted
or not.  This caused confusion since variables which were already secured did
not need to be re-encrypted.

This does _not_ solve the problem of the fact that Chamber re-encrypts already
encrypted values.  That fix is coming in a subsequent commit.