Feature: Add 'to_flattened_name_hash' to Settings
This method takes any hierarchical layout that a setting may have and returns
a one-dimensional `Hash` where the keys are the flattened components of the
setting's name hierarchy.

Example:

```ruby
Settings.new(settings: {
               my_setting: 'value',
               level_1:    {
                 level_2:    {
                   some_setting: 'hello',
                 },
                 body:       'gracias',
               },
             }).to_flattened_name_hash
  ['my_setting']                         => 'value',
  ['level_1', 'level_2', 'some_setting'] => 'hello',
  ['level_1', 'body']                    => 'gracias',
}
```

This could be used in multiple places such as displaying the settings via `to_s`
which is currently being run through `SystemEnvironment`, but it can also be
used for determining where in the YAML file that an encrypted setting needs to
go.