Refactor: Change pre and post filters for secure and insecure
Now that we have `TranslateSecureKeysFilter`, we should be able to use it in the
post filter and allow all filters to be able to use the translated `raw_data`,
this means that these methods should now be chainable:

Example:

```ruby
my_settings.insecure.secure
```

There is a difference between things that actually _filter_ settings and things
that _convert_ settings.  Actual filters should be in the `prefilter` list,
items that translate should be in the `postfilter` list.