Feature: When securing files, do not rewrite the entire file
Now, when running `chamber secure`, chamber will only rewrite the lines in the
settings files that correspond to the items being secured.

Additionally, any whitespace before the key, after the key and before the colon,
as well as after the colon and before the value, will be preserved.

This logic only concerns itself with the root key, I don't see this as being
a problem because if the value is the same, it doesn't really matter which
version of the encrypted string is put in its place, but I could see this as
having a chance of being a potential bug down the road.

The update to the `File#secure` method hasn't been refactored as much as I would
like, but I wanted to get it in the wild before I spent a bunch of time on it to
see if this is going to be a sustainable solution.

References #11