Feature: Allow any values (including complex ones) to be secured
Previously, if you attempted to secure an integer value such as `12345` or
something other than a `String`, chamber would silently ignore it and pass it
through unencrypted.

Obviously this is surprising to users considering if they put `_secure_` in
front of their setting name and run `chamber secure`, they would expect that
said setting would be secured.

The side effect of this change is that, in order to _decrypt_ values, we must
take into account users who have the old, unmarshalled values currently in their
YAML files.

In order to account for this, we're catching the `TypeError` which is raised if
`Marshal#load` cannot read the string and returning the unencrypted string
instead.

If it _does_ read it, then it loads it up and returns that.

--------------------------------------------------------------------------------
Branch:        feature/allow-any-values-to-be-encrypted
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>