Bugfix: Special Regex characters caused values to not encrypt
Due to the way we were looking for existing lines in a settings file, which was
basically interpolating the settings name/value into a `Regexp`, if the name or
value had any special regex characters such as '+', '\', etc, they were treated
as part of the regex instead of as literal characters.

Therefore, this settings file:

```yaml
my_setting:
  _secure_has_a_special+_char: "My value is special?"
```

would not properly encrypt.

Now we are escaping both the name and the value before putting it into the regex
and it works fine.

--------------------------------------------------------------------------------
Branch:        bug/names-and-values-with-regex-chars-do-not-encrypt
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>