Refactor: Begin extracting decryption key logic into its own class
There is a lot of key logic floating around the codebase.  I'd like to
consolidate it into one or two classes.  This is the first pass at that.

Additionally, the `DecryptionKey.resolve` will always return the _contents_ of
the key and not the filename.  It does this for consistency and to allow us to
change the behavior so that we can pull the key from environment variables as
well as files on disk.

--------------------------------------------------------------------------------
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>