Bugfix: Settings overridden by environment variables were decrypted
If a setting is set in an environment variable, it should not also be decrypted
from the YAML file.  The environment variable should take precedence.

Previously, the `TranslateSecureKeysFilter` ran before the `EnvironmentFilter`,
which meant that when the `EnvironmentFilter` got the settings, they no longer
had the `_secure_` prefix on them.

Unfortunately, when I reordered the filter list in `Settings` here: 96af235, the
`TranslateSecureKeysFilter` came _after_ the `EnvironmentFilter`.  This caused
the `EnvironmentFilter` to look for environment variable names such as
`MY_GROUP__SECURE_MY_SECURE_VALUE` rather than `MY_GROUP_MY_SECURE_VALUE`.

The ordering of the filters could not be reverted because the
`FailedDecryptionFilter` needs to come after the `EnvironmentFilter` but
_before_ the `TranslateSecureKeysFilter`.  This is due to the fact that, without
the `_secure_` on the setting name, the `FailedDecryptionFilter` wouldn't know
which settings it should throw errors on.

The solution was simply to remove `_secure_` from keys that `Environmentable`
was processing.

--------------------------------------------------------------------------------
Change-Id:     Ia1d24aa30b818909b31d8ce8ac87952c4fd9af10
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>