Fix: Thor throwing warnings about --shell option
This is due to a change in Thor from how it previously worked.
There isn't a huge incentive to passing 'shell' as an option so
we'll just add it to each individual call.

Not exactly dry, but it works.

Thor began throwing errors when the type of the argument passed
didn't match the type that it thought it should be.  Since
Thor is a command-line command generator, arguments cannot be
complex Ruby objects.  Therefore passing 'new' into it didn't
quite work.