Fix: Namespaces Passed As Hashes Not Being Transformed Properly
Why This Change Is Necessary
========================================================================

When a namespace was being passed in as a Hash (as is the default for
our Rails integration), the namespaces weren't being translated to an
array of namespaces, which is what the rest of the library expects.

This was working correctly for the CLI commands because they rely on the
fact that the system will pass through default values, but inside of the
server or the console, this was not working correctly.

Additionally, if a lambda or proc is passed in as a value, it was not
being `call`ed and therefore the value was not resolving.

How These Changes Address the Issue
========================================================================

We now convert hashes passed to the ContextResolver into an array of the
hash's values as well as call procs and lambdas if they are passed in.

Side Effects Caused By This Change
========================================================================

None known.

------------------------------------------------------------------------
Actions:
  * References #58