Change: Remove Dots And Dashes From Namespaces When Resolving Filenames
Why This Change Is Necessary
========================================================================

Previously dots and dashes were not removed from the namespaces when
looking up the filenames for keys.  This meant that when looking up
filenames for those namespaces, the per-namespace keys were not
resolving correctly.

How These Changes Address the Issue
========================================================================

This change removes dots and dashes from the namespaces when looking up
key filenames.

We also alter the script that generates key pairs so that it removes
dots and dashes from the filenames it generates.

Caveat
========================================================================

It's important to note that this _only_ applies to looking up the key
filename.  Inside the YAML files, dots and dashes are valid characters
and will be used when finding namespaced settings.

With a namespace of `example-host.com`, the private key filename would
be `.chamber.examplehostcom.pem`, however in the YAML file, you would
still do:

```yaml
example-host.com:
  my_setting: 'hello'
```

Side Effects Caused By This Change
========================================================================

Anyone relying on the erroneous functionality will break.

------------------------------------------------------------------------
Actions:
  * References #56