Implement the unsecure feature
Add instance unsecure method
which calls a new unsecure method on a FileSet
which is implemented as as decrypt method on a File instance

File#decrypt is a modified version of File#encrypt
I've also implemented new Settings filters to give the encrypted and decrypted versions. Not sure I understand this code well enough to know if there's a better way to have built it using the existing filters.